requirements.txt
streamlit
flask
openai
python-dotenv
pymupdf
docx2pdf